mage-T4/mage/cs4485_code/team2_package/requirements.txt
pandas>=1.5.0
psycopg2-binary>=2.9.0